feat: always use portfolio screen. Fix issue with account switching
