+++ packages/scaffold-ui/exports/index.ts
@@ -91,7 +91,6 @@ export * from '../src/partials/w3m-snackbar/index.js'
 export * from '../src/partials/w3m-alertbar/index.js'
 export * from '../src/partials/w3m-email-login-widget/index.js'
 export * from '../src/partials/w3m-account-default-widget/index.js'
-export * from '../src/partials/w3m-account-wallet-features-widget/index.js'
 export * from '../src/partials/w3m-account-activity-widget/index.js'
 export * from '../src/partials/w3m-account-nfts-widget/index.js'
 export * from '../src/partials/w3m-account-tokens-widget/index.js'
